Unify graph lenses onto one /graph surface; redirect /mesh and /context
Collapse the three separate lens routes into a single graph surface so
users switch lenses in place instead of navigating between parallel
pages. Attack Paths (/security-graph) and Asset Drift (/graph?scope=
asset-drift) keep their existing distinct targets; Lineage, Agent Mesh,
and Context are now all hosted on /graph and selected with ?lens=.

The Agent Mesh and Context lenses are distinct data pipelines (raw scan
mesh graph and the context-graph API) with their own panels, not filter
variants of the unified graph, so they are composed onto the surface
rather than folded into its filter model: their page bodies move to
components/mesh-lens-view.tsx and components/context-lens-view.tsx, and
app/graph/graph-surface.tsx lazily mounts the right one from the lens
param. All behavior is preserved.

The retired /mesh and /context routes become thin client-side redirects
to /graph?lens=mesh and /graph?lens=context, preserving query params so
old deep links never 404 — next.config redirects() cannot run under
output: export, so the redirect is a router.replace on mount (same
pattern already used by /vulns). Verified in the exported out/: both
routes emit real HTML whose page chunk carries the lens redirect.

The lens switcher now toggles the lens param on the unified surface for
Lineage/Mesh/Context, and nav/command-palette point Agent Mesh and
Context at the unified URLs. InsightLayerToggle exposes aria-pressed so
the active lens is programmatically verifiable.
