+++ ui/components/insight-layer-toggle.tsx
@@ -13,6 +13,7 @@ export function InsightLayerToggle({ layers, onToggle }: InsightLayerToggleProps
         <button
           key={layer.id}
           onClick={() => onToggle(layer.id)}
+          aria-pressed={layer.active}
           className={`flex max-w-full items-center gap-1 rounded-md border px-2 py-1 text-xs transition-colors ${
             layer.active
               ? "border-[color:var(--border-strong)] bg-[color:var(--surface-elevated)] text-[color:var(--foreground)]"
